undid betty changes to fix code errors i hate betty

+++ bit_manipulation/1-print_binary.c
@@ -8,7 +8,8 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int max = (uint) 1 << (sizeof(uint) * 8 - 1);
+	unsigned long int max = (unsigned long int) 1 << (
+		sizeof(unsigned long int) * 8 - 1);
 	int started = 0;
 
 	if (n == 0)
